+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * Fix #4998: removing the internal usage of the Serialization yaml mapper
 * Fix #5005: status operations use the context `subresource` setting
 * Fix #5022: adding additional buffering to ExecWatchInputStream
+* Fix #5052: add Quantity.fromNumericalAmount, the inverse of getNumericalAmount
 * Fix #5080: minimizing debug logs related to the backoff interval 
 
 #### Dependency Upgrade